feat(gfx_d3d): land upstream r_outdoor into POSIX/Switch build
Wire src/gfx_d3d/r_outdoor.cpp (179 lines) into both backends —
outdoor-floor heightmap pre-pass: trace-height into a CPU pic and
upload as a 2D luminance texture for terrain-aware shading.

Source fixes for -Werror clean build:
  - KISAKHACK uintptr_t bridge on the `(unsigned __int8 *)x` hex-rays
    artifact in Outdoor_ComputeTexels (`result` field used only for
    the loop-exit value).

Compat shim:
  - Add a D3DFORMAT enum shim (D3DFMT_UNKNOWN/A8/L8/A8L8/R32F) in
    kisak_compat.h, since multiple texture-format references in
    upstream renderer/asset code want these as integer constants.

Satellite stubs:
  - Image_Register / Image_Generate2D (live in r_image; nullptr/no-op
    until r_image lands).

+++ src/gfx_d3d/r_outdoor.cpp
@@ -115,7 +115,7 @@ unsigned __int8 *Outdoor_ComputeTexels()
         yWorld = outdoorGlob.invScale[1] * ((double)y + 0.5 - outdoorGlob.add[1]);
         for (x = 0; ; ++x)
         {
-            result = (unsigned __int8 *)x;
+            result = (unsigned __int8 *)(uintptr_t)x;
             if (x == outdoorMapSize[0])
                 break;
             xWorld = outdoorGlob.invScale[0] * ((double)x + 0.5 - outdoorGlob.add[0]);
